requirements.txt
adafruit_usb_host_descriptors==0.1.3
adafruit_displayio_ssd1306==2.1.1
adafruit_datetime==1.2.8
adafruit_midi==1.5.2
adafruit_bus_device==5.2.10
adafruit_display_text==3.1.4
